LPS-63381 move to module
