+++ modules/portal-reports-engine/portal-reporting-engine-api/src/main/java/com/liferay/portal/reports/engine/ContextClassloaderReportDesignRetriever.java
@@ -12,35 +12,35 @@
  * details.
  */
 
-package com.liferay.portal.kernel.bi.reporting;
-
-import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayInputStream;
+package com.liferay.portal.reports.engine;
 
 import java.io.InputStream;
+import java.io.Serializable;
 
 import java.util.Date;
 
 /**
  * @author Michael C. Han
  */
-public class MemoryReportDesignRetriever implements ReportDesignRetriever {
+public class ContextClassloaderReportDesignRetriever
+	implements ReportDesignRetriever, Serializable {
 
-	public MemoryReportDesignRetriever(
-		String reportName, Date modifiedDate, byte[] bytes) {
-
+	public ContextClassloaderReportDesignRetriever(String reportName) {
 		_reportName = reportName;
-		_modifiedDate = modifiedDate;
-		_bytes = bytes;
 	}
 
 	@Override
 	public InputStream getInputStream() {
-		return new UnsyncByteArrayInputStream(_bytes);
+		Thread currentThread = Thread.currentThread();
+
+		ClassLoader contextClassLoader = currentThread.getContextClassLoader();
+
+		return contextClassLoader.getResourceAsStream(_reportName);
 	}
 
 	@Override
 	public Date getModifiedDate() {
-		return _modifiedDate;
+		return new Date();
 	}
 
 	@Override
@@ -48,8 +48,6 @@ public class MemoryReportDesignRetriever implements ReportDesignRetriever {
 		return _reportName;
 	}
 
-	private final byte[] _bytes;
-	private final Date _modifiedDate;
 	private final String _reportName;
 
 }
