+++ modules/portal-reports-engine/portal-reporting-engine-api/src/main/java/com/liferay/portal/reports/engine/ReportDataSourceType.java
@@ -12,31 +12,28 @@
  * details.
  */
 
-package com.liferay.portal.kernel.bi.reporting;
+package com.liferay.portal.reports.engine;
 
 /**
- * @author Michael C. Han
+ * @author Gavin Wan
  */
-public enum ReportFormat {
+public enum ReportDataSourceType {
 
-	CSV("csv"), HTML("html"), PDF("pdf"), RTF("rtf"), TXT("txt"), XLS("xls"),
+	CSV("csv"), EMPTY("empty"), JDBC("jdbc"), PORTAL("portal"), XLS("xls"),
 	XML("xml");
 
-	public static ReportFormat parse(String value) {
+	public static ReportDataSourceType parse(String value) {
 		if (CSV.getValue().equals(value)) {
 			return CSV;
 		}
-		else if (HTML.getValue().equals(value)) {
-			return HTML;
+		else if (EMPTY.getValue().equals(value)) {
+			return EMPTY;
 		}
-		else if (PDF.getValue().equals(value)) {
-			return PDF;
+		else if (JDBC.getValue().equals(value)) {
+			return JDBC;
 		}
-		else if (RTF.getValue().equals(value)) {
-			return RTF;
-		}
-		else if (TXT.getValue().equals(value)) {
-			return TXT;
+		else if (PORTAL.getValue().equals(value)) {
+			return PORTAL;
 		}
 		else if (XLS.getValue().equals(value)) {
 			return XLS;
@@ -57,7 +54,7 @@ public enum ReportFormat {
 		return _value;
 	}
 
-	private ReportFormat(String value) {
+	private ReportDataSourceType(String value) {
 		_value = value;
 	}
