+++ modules/portal-reports-engine/portal-reporting-engine-api/src/main/java/com/liferay/portal/reports/engine/ReportDesignRetriever.java
@@ -12,28 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.kernel.bi.reporting;
+package com.liferay.portal.reports.engine;
 
-import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.resource.ResourceRetriever;
+
+import java.util.Date;
 
 /**
  * @author Michael C. Han
  */
-public class ReportGenerationException extends PortalException {
+public interface ReportDesignRetriever extends ResourceRetriever {
 
-	public ReportGenerationException() {
-	}
+	public Date getModifiedDate();
 
-	public ReportGenerationException(String msg) {
-		super(msg);
-	}
-
-	public ReportGenerationException(String msg, Throwable cause) {
-		super(msg, cause);
-	}
-
-	public ReportGenerationException(Throwable cause) {
-		super(cause);
-	}
+	public String getReportName();
 
 }
