+++ modules/portal-reports-engine/portal-reporting-engine-api/src/main/java/com/liferay/portal/reports/engine/ReportExportException.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.reports.engine;
+
+/**
+ * @author Michael C. Han
+ */
+public class ReportExportException extends ReportGenerationException {
+
+	public ReportExportException() {
+	}
+
+	public ReportExportException(String msg) {
+		super(msg);
+	}
+
+	public ReportExportException(String msg, Throwable cause) {
+		super(msg, cause);
+	}
+
+	public ReportExportException(Throwable cause) {
+		super(cause);
+	}
+
+}
